fix: satisfy backend-size limits for CI mergeability
Move factory session lifecycle handlers into handlers_factory_lifecycle.go
and extract pause/resume contract fixtures so backend-size verification passes.
