+++ pkg/api/contracttests/generated_contract_common_test.go
@@ -406,6 +406,7 @@ func generatedFactorySessionLifecycleEvents(t *testing.T, eventTime time.Time) [
 	}
 	events := make([]factoryapi.FactoryEvent, 0, 8)
 	events = append(events, generatedSessionLifecycleBracketEvents(t, fixture, eventTime)...)
+	events = append(events, generatedSessionLifecyclePauseResumeEvents(t, fixture, eventTime)...)
 	events = append(events, generatedSessionLifecycleOrchestratorEvents(t, fixture, eventTime)...)
 	events = append(events, generatedSessionLifecycleDispatchEvents(t, fixture, eventTime)...)
 	return events
@@ -439,42 +440,6 @@ func generatedSessionLifecycleBracketEvents(
 				PolicyHash: stringPtr("sha256:policy"),
 				ArgsDigest: stringPtr("sha256:args"),
 				StartedAt:  eventTime,
-			}),
-		},
-		{
-			SchemaVersion: factoryapi.AgentFactoryEventV1,
-			Id:            "event-session-paused",
-			Type:          factoryapi.FactoryEventTypeSessionPaused,
-			Context: factoryapi.FactoryEventContext{
-				Sequence:         11,
-				Tick:             6,
-				EventTime:        eventTime,
-				SessionId:        &fixture.sessionID,
-				SessionSequence:  intPtr(fixture.nextSessionSequence()),
-				OrchestratorKind: &fixture.orchestratorKind,
-				Source:           &fixture.source,
-			},
-			Payload: factoryEventPayload(t, factoryapi.SessionPausedEventPayload{
-				Status:   factoryapi.FactorySessionDurableLifecycleStatusPaused,
-				PausedAt: eventTime,
-			}),
-		},
-		{
-			SchemaVersion: factoryapi.AgentFactoryEventV1,
-			Id:            "event-session-resumed",
-			Type:          factoryapi.FactoryEventTypeSessionResumed,
-			Context: factoryapi.FactoryEventContext{
-				Sequence:         12,
-				Tick:             6,
-				EventTime:        eventTime,
-				SessionId:        &fixture.sessionID,
-				SessionSequence:  intPtr(fixture.nextSessionSequence()),
-				OrchestratorKind: &fixture.orchestratorKind,
-				Source:           &fixture.source,
-			},
-			Payload: factoryEventPayload(t, factoryapi.SessionResumedEventPayload{
-				Status:    factoryapi.FactorySessionDurableLifecycleStatusRunning,
-				ResumedAt: eventTime,
 			}),
 		},
 		{
@@ -521,6 +486,52 @@ func generatedSessionLifecycleBracketEvents(
 					Running:   0,
 					Completed: 2,
 				},
+			}),
+		},
+	}
+}
+
+func generatedSessionLifecyclePauseResumeEvents(
+	t *testing.T,
+	fixture *generatedSessionLifecycleFixture,
+	eventTime time.Time,
+) []factoryapi.FactoryEvent {
+	t.Helper()
+	return []factoryapi.FactoryEvent{
+		{
+			SchemaVersion: factoryapi.AgentFactoryEventV1,
+			Id:            "event-session-paused",
+			Type:          factoryapi.FactoryEventTypeSessionPaused,
+			Context: factoryapi.FactoryEventContext{
+				Sequence:         11,
+				Tick:             6,
+				EventTime:        eventTime,
+				SessionId:        &fixture.sessionID,
+				SessionSequence:  intPtr(fixture.nextSessionSequence()),
+				OrchestratorKind: &fixture.orchestratorKind,
+				Source:           &fixture.source,
+			},
+			Payload: factoryEventPayload(t, factoryapi.SessionPausedEventPayload{
+				Status:   factoryapi.FactorySessionDurableLifecycleStatusPaused,
+				PausedAt: eventTime,
+			}),
+		},
+		{
+			SchemaVersion: factoryapi.AgentFactoryEventV1,
+			Id:            "event-session-resumed",
+			Type:          factoryapi.FactoryEventTypeSessionResumed,
+			Context: factoryapi.FactoryEventContext{
+				Sequence:         12,
+				Tick:             6,
+				EventTime:        eventTime,
+				SessionId:        &fixture.sessionID,
+				SessionSequence:  intPtr(fixture.nextSessionSequence()),
+				OrchestratorKind: &fixture.orchestratorKind,
+				Source:           &fixture.source,
+			},
+			Payload: factoryEventPayload(t, factoryapi.SessionResumedEventPayload{
+				Status:    factoryapi.FactorySessionDurableLifecycleStatusRunning,
+				ResumedAt: eventTime,
 			}),
 		},
 	}
